ai_service/requirements.txt
flask
pandas
numpy
scikit-learn
joblib
psycopg2-binary
scipy
Pillow
piexif
